Move meta checks to the integrity module
