+++ src/integrity/mod.rs
@@ -2,6 +2,8 @@
 mod checks;
 pub mod error;
 mod issue;
+pub mod meta_check;
+mod meta_issue;
 mod repairer;
 mod utils;
 mod verifier;
